handle book picture for upload and delete

+++ ReadingHub/Cores/Repository/BookRepository.cs
@@ -4,24 +4,31 @@ using ReadingHub.Cores.Models;
 using ReadingHub.Cores.Validations.Exceptions;
 using ReadingHub.Persistence.Abstract;
 using ReadingHub.Persistence.Models;
+using ReadingHub.Unit.Abstracts;
 using ReadingHub.Unit.Abstracts.Repository;
 
+ 
 namespace ReadingHub.Cores.Repository
 {
     public class BookRepository : IBookRepository
     {
         private readonly IApplicationDbContext _context;
         public readonly IMapper _mapper;
+        private IHostEnvironment Environment;
+        private readonly IUserService _userService;
 
-        public BookRepository(IApplicationDbContext context,IMapper mapper)
+        public BookRepository(IUserService userService,IHostEnvironment env,IApplicationDbContext context,IMapper mapper)
         {
             _context = context;
             _mapper = mapper;  
+            Environment = env;
+            _userService = userService;
         }
        
 
         public async Task<int> PublishBook(BookViewModel model)
         {
+
             var book = _mapper.Map<BookViewModel,Book>(model);
 
              FileConvert(ref book,model.BookFile);
@@ -33,6 +40,8 @@ namespace ReadingHub.Cores.Repository
             {
                 return -1;
             }
+            SaveFileToDisk(model.Photo,createBook.Entity.Id);
+
             return createBook.Entity.Id;
         }
         public Task<bool> UpdateBook(BookViewModel model)
@@ -57,11 +66,22 @@ namespace ReadingHub.Cores.Repository
             _context.Complete(); 
             return Task.FromResult(true);
         }
+         async void SaveFileToDisk(IFormFile file,int identifier) {
+            string path = GetBookContentsDirectory();
+            Directory.CreateDirectory(path);
+            using (Stream fileStream = new FileStream(Path.Combine(path,identifier.ToString()+$".{file.FileName.Split('.')[1]}"), FileMode.Create))
+            {
+               await file.CopyToAsync(fileStream);
+            }
+        }
 
+        public string GetBookContentsDirectory() {
+            return Environment.ContentRootPath + "wwwroot/booksImgs";
+        }
           static void FileConvert(ref Book book,IFormFile file)
         {
             var dataStream = new MemoryStream();
-
+             
               file.CopyTo(dataStream);
             book.BookMimeType = file.ContentType;
 
@@ -74,6 +94,7 @@ namespace ReadingHub.Cores.Repository
         {
             var books = _context.Books.AsQueryable();
 
+            
             return Task.FromResult(_mapper.Map<IEnumerable<Book>,IEnumerable<GetBooksViewModel>>(books));
 
         }
@@ -90,15 +111,37 @@ namespace ReadingHub.Cores.Repository
 
             GuardException.NotFound(book, nameof(Book));
 
-            return Task.FromResult(_mapper.Map<Book, GetBookViewModel>(book));
+            var convert = _mapper.Map<Book, GetBookViewModel>(book);
+            convert.Photo =  "booksImgs/" + GetBookName(book.Id);
+            return Task.FromResult(convert);
         }
 
+        string GetBookName(int bookId) {
+           var files =  Directory.GetFiles(GetBookContentsDirectory());
+
+            foreach (var file in files)
+            {
+                
+                if (file.Contains(bookId.ToString())) { 
+                return bookId+"."+file.Split(".")[1];
+                }
+            }
+
+            return null;
+        }
         public Task<bool> DeleteBook(int bookId)
         {
+ 
             var book = _context.Books.FirstOrDefault(e => e.Id == bookId);
             GuardException.NotFound(book, nameof(Book));
+
+            if(book.AuthorId != _userService.GetUserId())
+                return Task.FromResult(false);
+
             _context.Books.Remove(book);
             _context.Complete();
+
+            File.Delete(Path.Combine(GetBookContentsDirectory(), GetBookName(book.Id)));
 
             return Task.FromResult(true);
         }
